Made the suggested changes
