+++ app/Validators/CreateUserValidator.ts
@@ -1,0 +1,13 @@
+import { schema, CustomMessages, rules } from '@ioc:Adonis/Core/Validator'
+import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
+
+export default class CreateUserValidator {
+  constructor(protected ctx: HttpContextContract) {
+  }
+
+  public schema = schema.create({
+    email: schema.string({}, [rules.email()]),
+    password: schema.string({}, [rules.minLength(8), rules.maxLength(16), rules.alphaNum()])
+  })
+  public messages: CustomMessages = {}
+}
